pyproject.toml
[build-system]
requires = ["setuptools>=40.8.0", "wheel", "setuptools_scm[toml]>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "PyFiberModes"
dynamic = ["version"]
description = "A package for light propagation in fiber optics."
license = {file = "LICENSE"}
readme = "README.rst"
requires-python = ">=3.10"
authors = [{ name="Martin Poinsinet de Sivry-Houle", email="martin.poinsinet.de.sivry@gmail.com"}]
classifiers = [
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: Implementation :: CPython',
    'Development Status :: 4 - Beta',
    'Topic :: Scientific/Engineering :: Physics',
    'License :: OSI Approved :: MIT License',
    'Intended Audience :: Science/Research',
]

keywords = ["finit-difference", "matrix"]

dependencies = [
    "PyOptik",
    "pyyaml ==6.0.1",
    "numpy ==1.*",
    "MPSPlots ~=1.5.1",
    "PyFinitDiff",
    "scipy ==1.*",
]

[tool.setuptools]
packages = {find = {where = ["."], include = ["PyFiberModes*"]}}
include-package-data = true


[tool.setuptools_scm]
write_to = "PyFiberModes/_version.py"
version_scheme = "only-version"
local_scheme = "no-local-version"

[project.optional-dependencies]
testing = [
    "pytest >=7.4,<9.0",
    "pytest-cov >=2,<6",
    "pytest-json-report ~=1.5",
    "coverage ~=7.6"
]


documentation = [
    "numpydoc ==1.6.0",
    "sphinx >=5.1.1",
    "sphinx-gallery ==0.15.0",
    "sphinx-rtd-theme ==2.0.0",
    "pydata-sphinx-theme ==0.14.1",
]

dev = [
    "flake8 ==7.1.1",
]

[tool.pytest.ini_options]
python_functions = "test*"
minversion = "6.0"
xfail_strict = true
log_cli_level = "INFO"
testpaths = [
    "tests/*",
]
addopts = [
    '-v',
    '-rA',
    '-rw',
    '-Werror',
    '--cov=PyFiberModes',
    '--cov-report=html',
    "--cov-report=term"
]


[tool.coverage.run]
source = ['PyFiberModes']
omit = [
    '*\__init__.py',
    '*\_version.py',
    '*\directories.py',
]
branch = true
relative_files = true